Moved unit tests for GetNoteIdsWithAtLeastKanjiInterval() into KanjiCardServiceTests.
